refactor: extract scroll diagnostics from MessageListScrollCoordinator into ScrollDiagnosticsRecorder (#21625)
